Replaced references to standard Java date functions with Joda time.
Since I was removing reliance on the standard Java date functions in
the Bishop library, I decided to the same with this sample
project.

In addition, I changed the way the last-modified date for the todo
group resource is calculated. There's now another reference that keeps
track of the latest modification date of the entire database.